+++ src/source-meta/index.js
@@ -1,0 +1,1 @@
+export { RowanSourceMeta } from "./source-meta.js";