+++ src/source-meta/source-meta.d.ts
@@ -1,0 +1,3 @@
+import "../elements.js";
+
+export { RowanSourceMeta } from "../../types/source-meta/source-meta.js";